internet-store-backend/requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.7
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
h11==0.14.0
httptools==0.6.4
idna==3.10
Markdown==3.7
msgpack==1.1.0
pillow==10.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz===2024.2
PyYAML==6.0.2
redis==5.2.0
sniffio==1.3.1
sqlparse==0.5.1
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
whitenoise==6.8.2
